requirements.txt
aiogram==3.0.0b6
aioredis==2.0.1
aiosqlite==0.18.0
python-dotenv==0.21.1
redis==4.4.2

